second commit
Proof of concept.
- Created objects for a Group and a Contact.
- Access phone's address book, choose a contact and add to a group.

TODO:
- Convert to pivot page UI.
- Allow the user to set reminders. Users may set a reminder of when to
send a message. Therefore a reminder has an associated message and
message type set at the time of a reminder. The reminder should be a
toast notification that will occur (even when the app is not running)
that directs directly to the send page. I.e., for texts the recipients
and message should already be entered. All fields can be preset but may
be optional.
- Database for storing groups of people. A few of a group's attributes
are scheduled reminders (a separate object, whose attributes are a
message and messaging type (SMS, Social networks, E-mail)), a collection
of contacts
- Allow users to set text reminders for a group
- Toast notifications reminding the user of scheduled times.